+++ R/qualification_level.R
@@ -1,10 +1,11 @@
 # See R/roxygen.R for documentation below that uses inline R code
 
-#' Identify undergraduate terms
+#' Categorize qualification level by term
 #' 
-#' Categorize terms as “undergrad” for terms before a student's first degree 
-#' or “post-bacc” (post-baccalaureate) for terms after the first degree  
-#' and add columns that support the findings. Post-baccalaureate terms are 
+#' Categorize the qualification level towards which a student is working in 
+#' each term. Two levels are used: “undergrad” for terms before a student's 
+#' first degree and “post-bacc” (post-baccalaureate) for terms after the first 
+#' degree. Added columns support the findings. Post-baccalaureate terms are 
 #' typically excluded from the `term, course,` and `degree` data tables.
 #'
 #' @param dframe `r dframe` with required variables `{mcid}` and one of
@@ -19,16 +20,16 @@
 #'   - `bacc` &nbsp;  Character. Term of a student's first
 #'      baccalaureate, encoded `YYYYT` or, if no degree recorded, `NA`.
 #'      Joined from the `term_degree` variable in `midf_table.`
-#'   - `focus` &nbsp;  Character, indicating that a term belongs
+#'   - `qual_level` &nbsp;  Character, indicating that a term belongs
 #'      to one of two groups: "undergrad" terms are those leading up to
 #'      and including the term in which a student completes their first
 #'      degree; and "post-bacc" (post-baccalaureate) for all terms after the
 #'      first degree.
 #' * `r not_preserved`
-#' @example man/examples/exa_undergrad_focus.R
+#' @example man/examples/exa_qualification_level.R
 #' @export
 #'
-undergrad_focus <- function(dframe, midf_table = degree) {
+qualification_level <- function(dframe, midf_table = degree) {
   #
   # ---------- initial assertions
 
@@ -45,11 +46,11 @@ undergrad_focus <- function(dframe, midf_table = degree) {
   # active column names
   reqd_dframe_vars <- c("mcid", term_var)
   reqd_table_vars <- c("mcid", "term_degree")
-  added_vars <- c("bacc", "focus")
+  added_vars <- c("bacc", "qual_level")
 
   # bind names for R CMD check
   bacc <- NULL
-  focus <- NULL
+  qual_level <- NULL
   IDX <- NULL
   TERM_VAR <- NULL
 
@@ -104,7 +105,7 @@ undergrad_focus <- function(dframe, midf_table = degree) {
   dframe <- DT[dframe, on = "mcid"]
 
   # assign term status labels
-  dframe[, focus := fifelse(
+  dframe[, qual_level := fifelse(
     TERM_VAR > bacc,
     "post-bacc",
     "undergrad",